+++ resources/views/examples/pages/home.veil.php
@@ -4,7 +4,7 @@
 
 <p>@say:common.welcome Your application is installed and ready to use.</p>
 
-<p>Bones v{{bones.version}}</p>
+<p>Bones v{{bones.version}} | App v{{webapp.version}}</p>
 
 @endsection