+++ resources/bootstrap.php
@@ -3,11 +3,6 @@
 use Bayfront\Bones\Application\Utilities\App;
 use Bayfront\Translation\Adapters\Local;
 use Bayfront\Translation\Translate;
-use Monolog\Formatter\LineFormatter;
-use Monolog\Handler\BrowserConsoleHandler;
-use Monolog\Handler\RotatingFileHandler;
-use Monolog\Level;
-use Monolog\Logger;
 
 /** @var Bayfront\Container\Container $container */
 
@@ -32,50 +27,17 @@ use Monolog\Logger;
 
 // Translate
 
-$container->set('Bayfront\Translation\Translate', function () {
+$adapter = new Local(App::resourcesPath('/translations'));
+$translate = new Translate($adapter, App::getConfig('webapp.locale.default', 'en'), true);
 
-    $adapter = new Local(App::resourcesPath('/translations'));
-
-    return new Translate($adapter, App::getConfig('app.locale.default', 'en'), true);
-
-});
-
+$container->set('Bayfront\Translation\Translate', $translate);
 $container->setAlias('translate', 'Bayfront\Translation\Translate');
 
-// Filesystem
+// WebApp service
 
-$container->set('League\Flysystem\Filesystem', function() {
+$webAppService = $container->make('Bayfront\BonesService\WebApp\WebAppService', [
+    'config' => (array)App::getConfig('webapp', [])
+]);
 
-    $adapter = new League\Flysystem\Local\LocalFilesystemAdapter(App::storagePath('/app'));
-    return new League\Flysystem\Filesystem($adapter);
-
-});
-
-$container->setAlias('filesystem', 'League\Flysystem\Filesystem');
-
-// Log
-
-$container->set('Monolog\Logger', function() {
-
-    $log = new Logger('app');
-
-    if (App::environment() == App::ENV_DEV) {
-        $log->pushHandler(new BrowserConsoleHandler());
-    }
-
-    $level = App::environment() == App::ENV_DEV ? Level::Debug : Level::Info;
-
-    $file_handler = new RotatingFileHandler(App::storagePath('/app/logs/app.log'), 90, $level, true, 0775);
-
-    $output = "[%datetime%] %channel%.%level_name%: %message% %context% %extra%\n";
-    $date_format = 'Y-m-d H:i:s T';
-
-    $file_handler->setFormatter(new LineFormatter($output, $date_format));
-
-    $log->pushHandler($file_handler);
-
-    return $log;
-
-});
-
-$container->setAlias('log', 'Monolog\Logger');
+$container->set('Bayfront\BonesService\WebApp\WebAppService', $webAppService);
+$container->setAlias('webAppService', 'Bayfront\BonesService\WebApp\WebAppService');