+++ resources/views/examples/layouts/container.veil.php
@@ -2,26 +2,21 @@
 /*
  * Layout: Page
  *
- * Sections:
+ * Places:
  *
- *   - content (required)
+ *  - content
+ *  - end_body (optional)
  *
- *   - head (optional)
- *   - end_body (optional)
- *
- * Predefined sections:
+ * Uses:
  *
  *   - examples/layouts/partials/head
  *   - examples/layouts/partials/header
  *   - examples/layouts/partials/footer
  *
- * $data array keys:
+ * Data:
  *
- *   - locale.all
+ *   - app.version
  *   - locale.current
- *   - page.title
- *   - page.description
- *   - year
  *
  */
 ?>
@@ -30,15 +25,15 @@
 
 @use:examples/layouts/partials/head
 
-<body class="tu-bg-default tu-text-default">
+<body class="bg-gray-100 dark:bg-gray-900 text-gray-800 dark:text-gray-100">
 
 <div id="content-wrap">
 
     @use:examples/layouts/partials/header
 
-    <main id="main" class="container xl:max-w-screen-xl mx-auto p-4 tu-bg-content">
+    <main id="main" class="container xl:max-w-screen-xl mx-auto rounded-md p-4 bg-white dark:bg-gray-800 shadow">
 
-        <div class="tu-typo text-center">
+        <div class="text-center">
 
             @place:content
 
@@ -51,6 +46,11 @@
 @use:examples/layouts/partials/footer
 
 <script src="@route:storage/assets/js/app.js"></script>
+
+<script>
+    let version = '{{app.version}}';
+    App.init(version);
+</script>
 
 ?@place:end_body