+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Removed Monolog library
 - Removed Flysystem library
+- Removed Scheduler service
 - Removed `app.version` config value
 - Removed `@route:` template tag filter