+++ CHANGELOG.md
@@ -12,6 +12,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities
 
+## [4.0.0] - 2025.01.10
+
+### Added
+
+- Added `bones-service-webapp`
+
+### Removed
+
+- Removed Monolog library
+- Removed Flysystem library
+- Removed Scheduler service
+- Removed `app.version` config value
+- Removed `@route:` template tag filter
+- Removed `@say:` template tag filter
+
 ## [3.1.0] - 2024.12.25
 
 ### Changed